Use Maybe (instead of an arbitrary MonadPlus) in decode and solveWith
As noted in #61, it's not clear what the intended behavior of
`decode`/`solveWith` should be for `MonadPlus` instances like `[]`. Moreover,
the internals of `ersatz` more-or-less already assume that we are working in
`Maybe`, and nearly all uses of `solveWith` in the wild pick `Maybe` to begin
with. To avoid potential confusion, let's just concretize `decode`/`solveWith`
to always work over `Maybe`.

Fixes #61.