Pass the right flags to depqbf based on its version
Version 6.03 of `depqbf` needs an additional flag (`--no-dynamic-nenofex`) that
older versions do not, so we now detect whether this flag is required by
parsing the output of `depqbf --version`.

As a result of this change, we no longer need to manually pass
`--no-dynamic-nenofex` to `depqbf` in the
`ersatz-{graph-coloring,circuit-synthesis}` examples.

Fixes #94.